FloatingWindowUnsigned_normal_implementation
Если при размере окна = 2,3 он ещё почти как справа-налево, хоть и медленнее, то при n=8 он раз в 5 медленнее становится

+++ src/algorithms/adelier/RightToLeftByAdelier.cpp
@@ -19,7 +19,7 @@ RightToLeft::~RightToLeft(){
 ZZ_p RightToLeft::exp(ZZ_p x, ZZ exponent){
 	ZZ_p y = conv<ZZ_p>(1);
 
-	long l = NumBits(exponent);
+	int l = NumBits(exponent);
 
 	for(int i = 0; i < l; ++i){
 		if (bit(exponent, i) == 1)
